+++ src/maybe.spec.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/ban-types */
 import { Maybe, Some } from './maybe';
 
 describe('Maybe', () => {
@@ -260,6 +261,30 @@ describe('Maybe', () => {
       it('None -> None', () => {
         const m1: Maybe<string> = Maybe.none;
         const m2: Maybe<string> = m1.notMatching(/i thh/);
+        expect(m2.isSome()).toBe(false);
+      });
+    });
+
+    describe('.match(...)', () => {
+      it('Some -> Some', () => {
+        const m1: Maybe<string> = Maybe.some('hi there');
+        const m2: Maybe<RegExpMatchArray> = m1.match(/i (th)/);
+        expect(m2.isSome()).toBe(true);
+        const r1 = m2.unwrap()[0];
+        const r2 = m2.unwrap()[1];
+        expect(r1).toBe('i th');
+        expect(r2).toBe('th');
+      });
+
+      it('Some -> None', () => {
+        const m1: Maybe<string> = Maybe.some('hi there');
+        const m2: Maybe<RegExpMatchArray> = m1.match(/no match/);
+        expect(m2.isSome()).toBe(false);
+      });
+
+      it('None -> None', () => {
+        const m1: Maybe<string> = Maybe.none;
+        const m2: Maybe<RegExpMatchArray> = m1.match(/no match/);
         expect(m2.isSome()).toBe(false);
       });
     });
@@ -553,6 +578,232 @@ describe('Maybe', () => {
           const maybe = Maybe.none.lte(BigInt(500));
           expect(maybe.isSome()).toEqual(false);
         });
+      });
+    });
+
+    describe('.at(...)', () => {
+      it('forward index', () => {
+        const m1: Maybe<number[]> = Maybe.some([1, 2, 3,]);
+
+        const r1: Maybe<number> = m1.at(0);
+        expect(r1.isSome()).toBe(true);
+        expect(r1.unwrap()).toBe(1);
+
+        const r2: Maybe<number> = m1.at(1);
+        expect(r2.isSome()).toBe(true);
+        expect(r2.unwrap()).toBe(2);
+
+        const r3: Maybe<number> = m1.at(2);
+        expect(r3.isSome()).toBe(true);
+        expect(r3.unwrap()).toBe(3);
+
+        const r4: Maybe<number> = m1.at(4);
+        expect(r4.isSome()).toBe(false);
+      });
+
+      it('reverse index', () => {
+        const m1: Maybe<number[]> = Maybe.some([1, 2, 3,]);
+
+        const r1: Maybe<number> = m1.at(-1);
+        expect(r1.isSome()).toBe(true);
+        expect(r1.unwrap()).toBe(3);
+
+        const r2: Maybe<number> = m1.at(-2);
+        expect(r2.isSome()).toBe(true);
+        expect(r2.unwrap()).toBe(2);
+
+        const r3: Maybe<number> = m1.at(-3);
+        expect(r3.isSome()).toBe(true);
+        expect(r3.unwrap()).toBe(1);
+
+        const r4: Maybe<number> = m1.at(-4);
+        expect(r4.isSome()).toBe(false);
+      });
+
+      describe('non-iterable', () => {
+        it('forward index', () => {
+          const m1: Maybe<{}> = Maybe.some({});
+
+          const r1: Maybe<{}> = m1.at(0);
+          expect(r1.isSome()).toEqual(true);
+          expect(r1.unwrap()).toEqual({});
+
+          const r2: Maybe<{}> = m1.at(1);
+          expect(r2.isSome()).toEqual(false);
+        });
+
+        it('reverse index', () => {
+          const m1: Maybe<{}> = Maybe.some({});
+
+          const r1: Maybe<{}> = m1.at(-1);
+          expect(r1.isSome()).toEqual(true);
+          expect(r1.unwrap()).toEqual({});
+
+          const r2: Maybe<{}> = m1.at(-2);
+          expect(r2.isSome()).toEqual(false);
+        });
+      });
+
+      describe('union[]', () => {
+        it('forward index', () => {
+          const m1: Maybe<(number | string | boolean)[]> = Maybe.some([1, '2', false,]);
+
+          const r1: Maybe<string | number | boolean> = m1.at(0);
+          expect(r1.isSome()).toBe(true);
+          expect(r1.unwrap()).toBe(1);
+
+          const r2: Maybe<string | number | boolean>  = m1.at(1);
+          expect(r2.isSome()).toBe(true);
+          expect(r2.unwrap()).toBe('2');
+
+          const r3: Maybe<string | number | boolean>  = m1.at(2);
+          expect(r3.isSome()).toBe(true);
+          expect(r3.unwrap()).toBe(false);
+
+          const r4: Maybe<string | number | boolean>  = m1.at(3);
+          expect(r4.isSome()).toBe(false);
+        });
+
+        it('reverse index', () => {
+          const m1: Maybe<(number | string | boolean)[]> = Maybe.some([1, '2', false,]);
+
+          const r1: Maybe<string | number | boolean> = m1.at(-1);
+          expect(r1.isSome()).toBe(true);
+          expect(r1.unwrap()).toBe(false);
+
+          const r2: Maybe<string | number | boolean>  = m1.at(-2);
+          expect(r2.isSome()).toBe(true);
+          expect(r2.unwrap()).toBe('2');
+
+          const r3: Maybe<string | number | boolean>  = m1.at(-3);
+          expect(r3.isSome()).toBe(true);
+          expect(r3.unwrap()).toBe(1);
+
+          const r4: Maybe<string | number | boolean>  = m1.at(-4);
+          expect(r4.isSome()).toBe(false);
+        });
+      });
+
+      describe('[tuple]', () => {
+        it('forward index', () => {
+          const m1: Maybe<[number, string, boolean]> = Maybe.some([1, '2', false,]);
+
+          const r1: Maybe<number> = m1.at(0);
+          expect(r1.isSome()).toBe(true);
+          expect(r1.unwrap()).toBe(1);
+
+          const r2: Maybe<string> = m1.at(1);
+          expect(r2.isSome()).toBe(true);
+          expect(r2.unwrap()).toBe('2');
+
+          const r3: Maybe<boolean> = m1.at(2);
+          expect(r3.isSome()).toBe(true);
+          expect(r3.unwrap()).toBe(false);
+
+          const r4: Maybe<undefined> = m1.at(3);
+          expect(r4.isSome()).toBe(false);
+        });
+
+        it('reverse index', () => {
+          const m1: Maybe<[number, string, boolean]> = Maybe.some([1, '2', false,]);
+
+          const r1: Maybe<string | number | boolean> = m1.at(-1);
+          expect(r1.isSome()).toBe(true);
+          expect(r1.unwrap()).toBe(false);
+
+          const r2: Maybe<string | number | boolean> = m1.at(-2);
+          expect(r2.isSome()).toBe(true);
+          expect(r2.unwrap()).toBe('2');
+
+          const r3: Maybe<string | number | boolean> = m1.at(-3);
+          expect(r3.isSome()).toBe(true);
+          expect(r3.unwrap()).toBe(1);
+
+          const r4: Maybe<string | number | boolean> = m1.at(-4);
+          expect(r4.isSome()).toBe(false);
+        });
+      });
+
+      describe('set', () => {
+        it('forward index', () => {
+          const m1: Maybe<Set<number>> = Maybe.some(new Set([1, 2, 3,]));
+
+          const r1: Maybe<number> = m1.at(0);
+          expect(r1.isSome()).toBe(true);
+          expect(r1.unwrap()).toBe(1);
+
+          const r2: Maybe<number> = m1.at(1);
+          expect(r2.isSome()).toBe(true);
+          expect(r2.unwrap()).toBe(2);
+
+          const r3: Maybe<number> = m1.at(2);
+          expect(r3.isSome()).toBe(true);
+          expect(r3.unwrap()).toBe(3);
+
+          const r4: Maybe<number> = m1.at(3);
+          expect(r4.isSome()).toBe(false);
+        });
+
+        it('reverse index', () => {
+          const m1: Maybe<Set<number>> = Maybe.some(new Set([1, 2, 3,]));
+
+          const r1: Maybe<number> = m1.at(-1);
+          expect(r1.isSome()).toBe(true);
+          expect(r1.unwrap()).toBe(3);
+
+          const r2: Maybe<number> = m1.at(-2);
+          expect(r2.isSome()).toBe(true);
+          expect(r2.unwrap()).toBe(2);
+
+          const r3: Maybe<number> = m1.at(-3);
+          expect(r3.isSome()).toBe(true);
+          expect(r3.unwrap()).toBe(1);
+
+          const r4: Maybe<number> = m1.at(-4);
+          expect(r4.isSome()).toBe(false);
+        });
+      });
+
+      describe('None -> None', () => {
+        it('forward index', () => {
+          const m1: Maybe<number> = Maybe.none;
+          const r1: Maybe<never> = m1.at(0);
+          expect(r1.isSome()).toBe(false);
+        });
+
+        it('reverse index', () => {
+          const m1: Maybe<number> = Maybe.none;
+          const r1: Maybe<never> = m1.at(-1);
+          expect(r1.isSome()).toBe(false);
+        });
+      });
+    });
+
+    describe('.pluck(...)', () => {
+      interface Pluckable { def: string; undef?: string }
+      describe('Some -> Some', () => {
+        it('required', () => {
+          const m1: Maybe<Pluckable> = Maybe.some({ def: 'hi', });
+          const m2: Maybe<string> = m1.pluck('def');
+          expect(m2.isSome()).toEqual(true);
+          expect(m2.unwrap()).toEqual('hi');
+        });
+        it('optional', () => {
+          const m1: Maybe<Pluckable> = Maybe.some({ def: 'hi', undef: undefined, });
+          const m2: Maybe<string | undefined> = m1.pluck('undef');
+          expect(m2.isSome()).toEqual(true);
+          expect(m2.unwrap()).toEqual(undefined);
+        });
+      });
+      it('Some -> None', () => {
+        const m1: Maybe<Pluckable> = Maybe.some({ def: 'hi', });
+        const m2: Maybe<string | undefined> = m1.pluck('undef');
+        expect(m2.isSome()).toEqual(false);
+      });
+      it('None -> None', () => {
+        const m1: Maybe<never>  = Maybe.none;
+        const m2: Maybe<never> = m1.pluck('undef');
+        expect(m2.isSome()).toEqual(false);
       });
     });
   });